Installed dependencies, tailwind and prettier, and fixed various bugs I noticed.
Tailwind is a styling tool to assist with the inline styling going forward. Heres a cheatsheet to get familiar with it, i recommend an extension for autofill tailwind so it will suggest the commands for you, https://tailwindcomponents.com/cheatsheet/
Prettier is autoformatting for our code, basically will enforce the tabbing policy and automatically synchronize with the plugin if you have it. Basically it will help keep our code clean and to es6 standards.